rename rule to be all camel-case
