+++ index.js
@@ -3,6 +3,5 @@
 module.exports = {
   rules: {
     'react-camel-case': require('./rules/react-camel-case'),
-    'react-svg-camel-case': require('./rules/react-svg-camel-case'),
   },
 }
